+++ projects/Mallard/src/navigation/index.ts
@@ -8,10 +8,8 @@ import { HomeScreen } from '../screens/home-screen'
 import { IssueScreen } from '../screens/issue-screen'
 import { ArticleScreen } from '../screens/article-screen'
 import { SettingsScreen } from '../screens/settings-screen'
-import { DownloadScreen } from '../screens/settings/download-screen'
-import { ApiScreen } from '../screens/settings/api-screen'
 import { color } from 'src/theme/color'
-import { Animated, Easing, View } from 'react-native'
+import { Animated, Easing } from 'react-native'
 import { useSettings } from 'src/hooks/use-settings'
 import {
     OnboardingIntroScreen,
@@ -19,11 +17,6 @@ import {
 } from 'src/screens/onboarding-screen'
 import { GdprConsentScreen } from 'src/screens/settings/gdpr-consent-screen'
 import { CasSignInScreen } from 'src/screens/settings/cas-sign-in-screen'
-import { TermsAndConditionsScreen } from 'src/screens/settings/terms-and-conditions-screen'
-import { HelpScreen } from 'src/screens/settings/help-screen'
-import { FAQScreen } from 'src/screens/settings/faq-screen'
-import { CreditsScreen } from 'src/screens/settings/credits-screen'
-import { PrivacyPolicyScreen } from 'src/screens/settings/privacy-policy-screen'
 import { NavigationScreenProp } from 'react-navigation'
 import { mapNavigationToProps } from './helpers'
 import { shouldShowOnboarding } from 'src/helpers/settings'
@@ -33,28 +26,7 @@ import {
 } from './interpolators'
 import { supportsTransparentCards } from 'src/helpers/features'
 import { AuthSwitcherScreen } from 'src/screens/identity-login-screen'
-
-const routeNames = {
-    Issue: 'Issue',
-    Article: 'Article',
-    IssueList: 'IssueList',
-    Downloads: 'Downloads',
-    Settings: 'Settings',
-    Endpoints: 'Endpoints',
-    GdprConsent: 'GdprConsent',
-    PrivacyPolicy: 'PrivacyPolicy',
-    TermsAndConditions: 'TermsAndConditions',
-    Help: 'Help',
-    Credits: 'Credits',
-    FAQ: 'FAQ',
-    SignIn: 'SignIn',
-    CasSignIn: 'CasSignIn',
-    onboarding: {
-        OnboardingStart: 'OnboardingStart',
-        OnboardingConsent: 'OnboardingConsent',
-        OnboardingConsentInline: 'OnboardingConsentInline',
-    },
-}
+import { routeNames } from './routes'
 
 const navOptionsWithGraunHeader = {
     headerStyle: {
