+++ projects/Mallard/src/navigation/routes.ts
@@ -1,0 +1,21 @@
+export const routeNames = {
+    Issue: 'Issue',
+    Article: 'Article',
+    IssueList: 'IssueList',
+    Downloads: 'Downloads',
+    Settings: 'Settings',
+    Endpoints: 'Endpoints',
+    GdprConsent: 'GdprConsent',
+    PrivacyPolicy: 'PrivacyPolicy',
+    TermsAndConditions: 'TermsAndConditions',
+    Help: 'Help',
+    Credits: 'Credits',
+    FAQ: 'FAQ',
+    SignIn: 'SignIn',
+    CasSignIn: 'CasSignIn',
+    onboarding: {
+        OnboardingStart: 'OnboardingStart',
+        OnboardingConsent: 'OnboardingConsent',
+        OnboardingConsentInline: 'OnboardingConsentInline',
+    },
+}
